feat : removeCartItem - authorization 추가

+++ controller/cart-controller.js
@@ -71,17 +71,30 @@ const getCartItems = (req, res) => {
 
 // 장바구니 아이템 삭제
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id;
 
-    let sql = "DELETE FROM cartItems WHERE id = ?;"
-    conn.query(sql, cartItemId,
-        (err, results) => {
-            if(err) {
-                console.log(err);
-                return res.status(StatusCodes.BAD_REQUEST).end();
-            }
-            return res.status(StatusCodes.OK).json(results);
-    })
+    let authorization = ensureAuthorization(req, res);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인하세요"
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const cartItemId = req.params.id;
+
+        let sql = "DELETE FROM cartItems WHERE id = ?;"
+        conn.query(sql, cartItemId,
+            (err, results) => {
+                if(err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
 }
 
 module.exports = {
